updated playerminiwindow to be sized only according to the main region (not the included regions)

+++ Classes/Player.h
@@ -72,7 +72,7 @@
 -(void) IncreasQuestionsPassed:(Question*) question;
 //-(void) IncreaseRoundScoreBy:(NSInteger) value;
 -(void) IncreaseScoreBy:(NSInteger) value;
--(NSInteger) GetQuestionsPassed;
+-(long) GetQuestionsPassed;
 -(NSInteger) GetLastRoundScore;
 -(void) SetPositionByScore:(NSInteger) position;
